+++ discO/utils/__init__.py
@@ -11,77 +11,11 @@ __all__ = [
 ##############################################################################
 # IMPORTS
 
-# BUILT-IN
-import typing as T
+from ._framelike import resolve_framelike
+from . import vectorfield
+from .vectorfield import *  # noqa: F401, F403
 
-# THIRD PARTY
-from astropy.coordinates import (
-    BaseCoordinateFrame,
-    SkyCoord,
-    sky_coordinate_parsers,
-)
-
-# PROJECT-SPECIFIC
-from discO.config import conf
-from discO.type_hints import FrameLikeType
-
-##############################################################################
-# CODE
-##############################################################################
-
-
-def resolve_framelike(
-    frame: T.Optional[FrameLikeType],
-    error_if_not_type: bool = True,
-):
-    """Determine the frame and return a blank instance.
-
-    .. todo::
-
-        should this be in skycoord.utils? (but conf)
-
-        This as a decorator, so it autoresolves
-
-    Parameters
-    ----------
-    frame : frame-like instance or None (optional)
-        If BaseCoordinateFrame, replicates without data.
-        If str, uses astropy parsers to determine frame class
-        If None (default), gets default frame name from config, and parses.
-
-    error_if_not_type : bool
-        Whether to raise TypeError if `frame` is not one of the allowed types.
-
-    Returns
-    -------
-    frame : `~astropy.coordinates.BaseCoordinateFrame` instance
-        Replicated without data.
-
-    """
-    # If no frame is specified, assume that the input footprint is in a
-    # frame specified in the configuration
-    if frame is None:
-        frame: str = conf.default_frame
-
-    if isinstance(frame, str):
-        frame = sky_coordinate_parsers._get_frame_class(frame.lower())()
-    elif isinstance(frame, BaseCoordinateFrame):
-        frame = frame.replicate_without_data()
-    elif isinstance(frame, SkyCoord):
-        frame = frame.frame.replicate_without_data()
-
-    elif error_if_not_type:
-        raise TypeError(
-            "Input coordinate frame must be an astropy "
-            "coordinates frame subclass *instance*, not a "
-            "'{}'".format(frame.__class__.__name__),
-        )
-
-    return frame
-
-
-# /def
-
+__all__ += vectorfield.__all__
 
 ##############################################################################
 # END
